+++ eyes.sdk.core/src/test/java/com/applitools/eyes/visualgrid/services/TestRunningTest.java
@@ -23,7 +23,7 @@ public class TestRunningTest extends ReportingTestSuite {
 
     @BeforeMethod
     public void beforeEach() {
-        runningTest = new VisualGridRunningTest(mock(RenderBrowserInfo.class), new Logger(), new Configuration());
+        runningTest = new VisualGridRunningTest(new Logger(), "", mock(RenderBrowserInfo.class), new Configuration());
         runningTest.issueCheck(mock(ICheckSettings.class), mock(List.class), "");
         runningTest.issueCheck(mock(ICheckSettings.class), mock(List.class), "");
         runningTest.issueCheck(mock(ICheckSettings.class), mock(List.class), "");